refactor(HttpError): create httpError util and import them inside every file that uses it(#37)
