+++ src/controllers/authController.ts
@@ -17,15 +17,7 @@ import {
 
 import { prisma } from "../index.js";
 
-class HttpError extends Error {
-  statusCode: number;
-
-  constructor(message: string, statusCode: number) {
-    super(message);
-    this.statusCode = statusCode;
-    this.name = "HttpError";
-  }
-}
+import HttpError from "../utils/httpError.js";
 
 interface AuthenticatedRequest extends Request {
   cookies: {
