+++ src/utils/httpError.ts
@@ -1,0 +1,11 @@
+class HttpError extends Error {
+  statusCode: number;
+
+  constructor(message: string, statusCode: number) {
+    super(message);
+    this.statusCode = statusCode;
+    this.name = "HttpError";
+  }
+}
+
+export default HttpError;
